queue/requirements.txt
celery
django-async-messages
